move action files to action package; make test pass by switching to expect double
