+++ core/src/main/java/org/javarosa/core/model/actions/ActionController.java
@@ -1,11 +1,13 @@
-package org.javarosa.core.model;
+package org.javarosa.core.model.actions;
 
+
+import org.javarosa.core.model.FormDef;
 import org.javarosa.core.model.instance.TreeReference;
-import org.javarosa.core.services.storage.Persistable;
 import org.javarosa.core.util.externalizable.DeserializationException;
 import org.javarosa.core.util.externalizable.ExtUtil;
 import org.javarosa.core.util.externalizable.ExtWrapListPoly;
 import org.javarosa.core.util.externalizable.ExtWrapMap;
+import org.javarosa.core.util.externalizable.Externalizable;
 import org.javarosa.core.util.externalizable.PrototypeFactory;
 
 import java.io.DataInputStream;
@@ -18,14 +20,12 @@ import java.util.Vector;
  * Registers actions that should be triggered by certain events, and handles the triggering
  * of those actions when an event occurs.
  *
- * Created by amstone326 on 2/2/16.
+ * @author Aliza Stone
  */
-public class ActionController implements Persistable {
-
-    private int id;
+public class ActionController implements Externalizable {
 
     // map from an event to the actions it should trigger
-    private Hashtable<String, Vector<Action>> eventListeners = new Hashtable<String, Vector<Action>>();
+    private Hashtable<String, Vector<Action>> eventListeners = new Hashtable<String, Vector<org.javarosa.core.model.actions.Action>>();
 
     public Vector<Action> getListenersForEvent(String event) {
         if (this.eventListeners.containsKey(event)) {
@@ -60,18 +60,8 @@ public class ActionController implements Persistable {
     }
 
     @Override
-    public int getID() {
-        return id;
-    }
-
-    @Override
-    public void setID(int id) {
-        this.id = id;
-    }
-
-    @Override
     public void readExternal(DataInputStream inStream, PrototypeFactory pf) throws IOException, DeserializationException {
-        eventListeners = (Hashtable<String, Vector<Action>>) ExtUtil.read(inStream,
+        eventListeners = (Hashtable<String, Vector<org.javarosa.core.model.actions.Action>>) ExtUtil.read(inStream,
                 new ExtWrapMap(String.class, new ExtWrapListPoly()), pf);
     }
