+++ core/src/main/java/org/javarosa/form/api/FormEntryController.java
@@ -16,7 +16,7 @@
 
 package org.javarosa.form.api;
 
-import org.javarosa.core.model.Action;
+import org.javarosa.core.model.actions.Action;
 import org.javarosa.core.model.FormIndex;
 import org.javarosa.core.model.QuestionDef;
 import org.javarosa.core.model.data.IAnswerData;
